requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.36.4
google-cloud-secret-manager==2.16.4
python-gitlab==3.15.0
pydantic==2.5.2
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-generativeai==0.3.2
requests==2.31.0
PyYAML==6.0.1
google-cloud-run==0.10.0
google-cloud-artifact-registry==1.8.0 